Moved crate_universe `TreeResolver` into it's own file.
